Add files via upload
The mean squared error (MSE) is 4.950, which means that on average, the predicted values are off by about 2 rings. The mean absolute error (MAE) is 1.607, which means that on average, the predicted values are off by about 1-2 rings.

The R-squared score is 0.543, which indicates that approximately 54.3% of the variation in the dependent variable (rings) can be explained by the independent variables (sex, length, diameter, height, weight, and type). This is a moderate level of predictive power, but it still leaves a significant amount of unexplained variation.

Overall, the linear regression model using the given independent variables has some predictive power, but it may not be sufficient for accurate predictions in all cases. Further analysis and modeling may be necessary to improve the accuracy of predictions.